delete unused relations and rename relations hpp
