+++ cpp/src/barretenberg/honk/composer/ultra_honk_composer.test.cpp
@@ -1,7 +1,7 @@
 // #include "ultra_honk_composer.hpp"
 // #include "barretenberg/common/log.hpp"
 // #include "barretenberg/honk/proof_system/ultra_prover.hpp"
-// #include "barretenberg/honk/sumcheck/relations/relation.hpp"
+// #include "barretenberg/honk/sumcheck/relations/relation_parameters.hpp"
 // #include "barretenberg/numeric/uint256/uint256.hpp"
 // #include <cstddef>
 // #include <cstdint>
