+++ cpp/src/barretenberg/honk/sumcheck/relations/gen_perm_sort_relation.hpp
@@ -3,7 +3,7 @@
 #include <tuple>
 
 #include "../polynomials/univariate.hpp"
-#include "relation.hpp"
+#include "relation_parameters.hpp"
 
 namespace proof_system::honk::sumcheck {
