+++ cpp/src/barretenberg/honk/sumcheck/relations/relation_consistency.test.cpp
@@ -1,8 +1,7 @@
 #include "barretenberg/honk/sumcheck/relations/lookup_relation.hpp"
 #include "barretenberg/honk/sumcheck/relations/ultra_arithmetic_relation.hpp"
-#include "barretenberg/honk/sumcheck/relations/ultra_arithmetic_relation_secondary.hpp"
 #include "barretenberg/honk/flavor/standard.hpp"
-#include "relation.hpp"
+#include "relation_parameters.hpp"
 #include "arithmetic_relation.hpp"
 #include "permutation_relation.hpp"
 #include "../polynomials/univariate.hpp"
