+++ cpp/src/barretenberg/honk/sumcheck/sumcheck.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "barretenberg/common/serialize.hpp"
 #include <array>
-#include "barretenberg/honk/sumcheck/relations/relation.hpp"
+#include "barretenberg/honk/sumcheck/relations/relation_parameters.hpp"
 #include "barretenberg/honk/transcript/transcript.hpp"
 #include "barretenberg/honk/utils/grand_product_delta.hpp"
 #include "barretenberg/common/throw_or_abort.hpp"
